0. Тестирование стороны прямоугольника

+++ src/ru/job4j/condition/SqArea.java
@@ -1,13 +1,15 @@
 package ru.job4j.condition;
 
 public class SqArea {
-    public static double square(int p, int k){
-        int one = p/(2*(k+1));
+
+    public static int square(int p, int k) {
+        int one = p / (2 * (k + 1));
         int two = one * k;
-        double rsl = one * two;
+        int rsl = one * two;
         return rsl;
     }
-    public static void main(String[] args){
+
+    public static void main(String[] args) {
         double result1 = SqArea.square(6, 2);
         System.out.println("p = 6, k = 2, s = 2, real = " + result1);
     }
